+++ src/App.css
@@ -5,7 +5,7 @@ background-color: #282c34;
 
 .App {
 text-align: center;
-max-width: 728px;
+max-width: 2400px;
 margin: 0 auto;
 }
 
@@ -16,7 +16,7 @@ min-height: 50px;
 color: white;
 position: fixed;
 width: 100%;
-max-width: 728px;
+max-width: 2400px;
 top: 0;
 display: flex;
 align-items: center;
@@ -61,13 +61,13 @@ position: fixed;
 bottom: 0;
 background-color: rgb(24, 23, 23);
 width: 100%;
-max-width: 728px;
+max-width: 2400px;
 display: flex;
 font-size: 1.5rem;
 }
 
 form button {
-width: 20%;
+width: 10%;
 background-color: rgb(56, 56, 143);
 }
 
@@ -81,10 +81,11 @@ color: white;
 outline: none;
 border: none;
 padding: 0 10px;
+
 }
 
 button {
-background-color: #282c34; /* Green */
+background-color: #C4C4C4;
 border: none;
 color: white;
 padding: 15px 32px;
@@ -94,6 +95,7 @@ display: inline-block;
 cursor: pointer;
 font-size: 1.25rem;
 }
+
 
 button:disabled {
 opacity: 0.5;
@@ -110,14 +112,13 @@ margin: 0 auto;
 
 
 
-
 ul, li {
 text-align: left;
 list-style: none;
 }
 
 p {
-max-width: 500px;
+max-width: 800px;
 margin-bottom: 12px;
 line-height: 24px;
 padding: 10px 20px;
@@ -125,6 +126,7 @@ border-radius: 25px;
 position: relative;
 color: white;
 text-align: center;
+font-size: 30px;
 }
 
 .message {
@@ -148,8 +150,8 @@ color: black;
 }
 
 img {
-width: 40px;
-height: 40px;
+width: 140px;
+height: 140px;
 border-radius: 50%;
 margin: 2px 5px;
 }
