+++ src/App.js
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import { Route, Switch } from 'react-router-dom';
 import HomePage from './Pages/HomePage.js';
 import NextPage from './Pages/NextPage.js';
-
 class App extends Component {
     render() {
         return (
