+++ src/style/ChatMessage.css
@@ -1,0 +1,34 @@
+.chat-message {
+    list-style-type: none;
+    display: inline-flex;
+    margin-bottom: 10px;
+    align-content: center;
+}
+.chat-profilebox {
+    justify-content: center;
+    text-align: center;
+    margin-right: 30px;
+    padding: auto;
+}
+.chat-profilebox > p {
+    font-size: 0.8em;
+    font-weight: bold;
+    margin-top: 5px;
+    justify-content: center;
+}
+.chat-actionbox {
+    display: flex;
+    padding: auto;
+    height: 50px;
+    text-align: center;
+    justify-content: center;
+    font-weight: bold;
+}
+.chat-scriptbox {
+    display: flex;
+    padding: 15px 15px 0px 15px;
+    height: 70px;
+    justify-content: center;
+    border-radius: 20%;
+    border: 1px solid #999;
+}
