+++ src/style/Modal.css
@@ -1,0 +1,32 @@
+.primary-btn{
+background-color: #4f4f4f;
+border-color: #4f4f4f;
+}
+
+.secondary-btn{
+background-color: #b2b2b2;
+border-color: #b2b2b2;
+color: black;
+}
+
+.profile-pic{
+border-radius: 40px;
+margin-left: 20px;
+margin-right: 15px;
+}
+
+.header-modalbutton{
+    border: 0px;
+    background: rgb(38, 51, 67);
+    font-size: 24px;
+    margin-left: 30px;
+    color: #007bff
+}
+
+.genre{
+margin: 10px;
+}
+
+.people-number{
+margin: 10px;
+}
