+++ src/style/WritingroomPage.css
@@ -1,0 +1,105 @@
+.writingroompage {
+    height: 90%;
+    width: 1000px;
+    margin: 0;
+}
+.writingroom-title {
+    width: 100%;
+    display: inline;
+    text-align: center;
+    padding: 10px 0 0 10px;
+    background-color: #6c757d;
+}
+.WR-title {
+    font-weight: 700;
+}
+.WR-backtoMain {
+    float: left;
+    color: white;
+    background-color: #6c757d;
+    border: none;
+    padding: 3px;
+    margin-left: 5px;
+}
+.WR-link {
+    color: white;
+}
+.WRdropdown-here {
+    background-color: #6c757d;
+    text-align: right;
+}
+.WRdropdownedMenu {
+    font-size: 0.9em;
+}
+.dropdown-toggle::after {
+    display:none;
+}
+.trytohandle {
+    height: 100%;
+}
+.WR-MainArea {
+    max-height: 530px;
+    background-color: 1px solid #f1f1f1;
+    padding: 20px;
+    margin: 0;
+    overflow: auto;
+}
+.WR-sidebar {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+    background-color: #f1f1f1;
+    position: fixed;
+    height: 100%;
+    overflow: auto;
+    float: right;
+  }
+.Wr-sidebar-comp {
+    display: block;
+    color: #000;
+    padding: 6px;
+    text-decoration: none;
+    text-align: center;
+    font-size: 2.2em;
+  }
+.Wr-sidebar-comp > .WR-small {
+    font-size: 16px;
+}  
+.Wr-sidebar-comp:hover {
+    background-color: #555;
+    color: white;
+}
+.WR-scriptbar {
+    display: flex;
+    align-items: center;
+    position: absolute;
+    bottom: 30px;
+    width: 62.5%;
+}
+#userCharacterImg {
+    height: 50px;
+    width: 50px;
+    border-radius: 50%;
+}
+#characterSelect {
+    background-color: transparent;
+    border: none;
+    border-radius: 50%;
+}
+.scriptInput {
+    padding: 10px;
+    width: 80%;
+    border-radius: 8px;
+    border-color: gray;
+    margin-left: 6px;
+    margin-right: 15px;
+}
+.WR-submitBtn > button {
+    padding: 10px;
+    background-color: white;
+    border-radius: 8px;
+    border: solid 1px gray;
+}
+.NV-li {
+    float:right;
+}
