+++ src/style/login.css
@@ -1,0 +1,32 @@
+.content {
+    margin-top: 50px;
+    border-radius: 20px;
+    padding: 50px 50px 50px 50px;
+    border: solid 1px gray;
+}
+.login-logo {
+    font-size: 2em;
+    font-weight: 900;
+    color: #43afc7;
+}
+.formInput {
+    border-radius: 8px;
+    border: solid 1px gray;
+    margin-left: 10px;
+    margin-bottom: 10px;
+    padding: 4px 15px 4px 15px;
+    font-family: Roboto;
+    font-size: 0.8em;
+    width: 200px;
+}
+.submitBtn {
+    border-radius: 10px;
+    border: none;
+    background-color: #43afc7;
+    padding: 10px 30px 10px 30px;
+    margin-left: 170px;
+}
+
+.linki {
+    font-size: 1em;
+}
